batch-processing/requirements.txt
python-dotenv
pymysql
numpy